Removed the files for the deprecated scanner.
Moved client specific .cmake files to /client directory.

Added wolfssl as external dependency for cryptography.

Added post build dependencies that copies the .so/.dll files to the same directory as the final executable.

Added post build dependencies that makes sure that .a/.lib files are correctly placed in the /lib directory with the rest of the library binaries.