Template-ified CHash so it now hashes a const_buffer instead of only a string_view.
Added expected_like concept to Characteristic.hpp (should maybe live in common.hpp?).

windows/CDevice's make function now return expected, in order to properly handle if make is called with an invalid address.

Fixed bug in windows/CDevice's make function that caused the if statement to not properly check if a BluetoothLEDevice was connected to successfully.

 Added header files to CmakeLists.txt sources - visual studio requires them otherwise it will not show the header files as part of the project.