Made current_mac_address() in ble_common.hpp return the mac address as capital letters for HEX values.
Added helper function in CAuthenticator.cpp to get the size of the expected hash type.

Fixed bug in CAuthenticator.cpp that caused the recieved hash to be hashed again.

Fixed bug in CAuthenticator.cpp that caused it to initialize spans into the buffer with the wrong sizes.

Added a constructor to CHash that takes two iterators to a span which allows CHasht to be created with a existing hash value.

Commented out whereami in CNimble.